#11 - fix: update of the edit profile

+++ database/migrations/2025_10_09_182257_create_users_table.php
@@ -18,7 +18,7 @@ return new class extends Migration {
             $table->string('password');
             $table->string('display_name')->nullable();
             $table->text('biography')->nullable();
-            $table->integer('telephone_number')->nullable();
+            $table->bigInteger('telephone_number')->nullable();
             $table->integer('prefix_telephone_number')->nullable();
             $table->string('avatar_img')->nullable();
             $table->integer('status');
